Fix unique_id for VRF
In the **VRF module**, the `self._mac_addr` is shared across all instances, which resulted in duplicate `self._unique_id` values.

Note that for non VRF, the _sub_mac_addr and _mac_addr are the same. this helps to simplfiy the code, as there is no need to distinguish between different cases.